Change type of array value from integer to single-precision floating point

+++ gcalc.c
@@ -3,12 +3,12 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-extern void scaling(int, int*, int);
+extern void scaling(int, float*, int);
 #ifdef __cplusplus
 }
 #endif
 
-void scaling(int num, int* arr, int size) {
+void scaling(int num, float* arr, int size) {
     int i;
     for(i = 0; i < size; i++) {
         arr[i] = arr[i] * num;
