+++ gcalc.f90
@@ -1,7 +1,7 @@
 subroutine scaling(num, arr, size_)
   integer, intent(in) :: num
   integer, intent(in) :: size_
-  integer, intent(inout) :: arr(size_)
+  real(4), intent(inout) :: arr(size_)
   integer i
 
   do i = 1, size_
